refactor(inspections-api): extract bulk sub-router
Dashboard, list, inspectors, bulk update, counts, schedule + sync conflicts
(8 routes) move to server/api/inspections/bulk.ts. Handler bodies byte-identical;
external paths and OpenAPI/RPC surface unchanged.
